FunLang. Add type checking of return statements in do blocks. Rather pervasive.
Change syntax of return statements to include monad annotation (in
order not to infer types).
Restructure type transformation (to be able to check whether a type
has a specific kind).
Simplify type checking success tests. Now we just check if the type
checker accepts a program. It was more like parsing tests before, which
is very time consuming and rather useless.