OOLang and MIL. New monad stacks: Pure_M and Impure_M. Improved type checking of bind in MIL.
OOLang: now it has Pure_M monad for Pure computations, which may not terminate
and may throw an exception. Impure_M effects ordering is also changed
(Lift and State) to match the state preservation intention.
Pure_M is a prefix of Impure_M. Lifts are not needed in the examples due to
the new bind checking in MIL.

MIL: Now bind type checking performs monad compatibility checking (alpha
equivalence or prefix) and chooses the "highest effects" type (among binding
expression and body expression) as a type of the whole expression.