FunLang. Pretty much a new code generator. Fully monadic code.
There is a lot of changes. Mainly, Pure_M monad, tricky type conversions,
data constructor wrappers, new tests. There are some doubts in certain parts
of the code, they also require more comments and tests, probably. But this
is a major step with FunLang code generation for now.
Some of the tests are commented out.