FunLang and MIL. New approach to code generation for built-in functions.
Now they have a special treatment in the code generation. But most
importantly, there is a problem with types for read functions in MIL
(they can throw an exception, see more in Notes). The solution for now is
to supply these functions and their types from the source language using
its monads (like IO_M, for example).