FunLang. Add polymorphic types code generation. Rewrite constructor wrappers generation.
Now it is more or less try to do the same things as type conversion but on
its own. Original constructor type is taken and not the wrapper's type.
This was done to deal with monads, because otherwise it is probably impossible
to distinguish between different types of type application.
Old code is kept in comments to save for history.