OOLang. Fix a bug with assignments.
First, the previousVar and currentVar did the wrong thing when going to
variable number 0 (which is the original variable).
Second, and more important is that the pre/post steps in the original
places did not work for statements which assigned an expression which referenced
a variable to that same variable. Those moved into the codeGenAssignMut
to be exactly around the target variable generation. In the codeGenStmts
pre/post steps are now done around the remaining statements only.